In this modification:  After retrieving the results from DuckDuckGo, we check if the number of results is less than 20. If the number of results is less than 20, we raise an exception to trigger a retry attempt. The script will then retry fetching the results according to the specified number of retry attempts and delay.
After retrieving the results from DuckDuckGo, we check if the number of results is less than 20.
If the number of results is less than 20, we raise an exception to trigger a retry attempt.
The script will then retry fetching the results according to the specified number of retry attempts and delay.